Deep_Learning_Nanodegree_Program/05_Teach_a_Quadcopter_How_to_Fly/quad_controller_rl/requirements_final.txt
absl-py==0.1.9
actionlib==1.11.9
angles==1.9.11
attrs==15.2.0
autobahn==0.10.3
backports.weakref==1.0.post1
beautifulsoup4==4.4.1
bleach==1.5.0
bondpy==1.7.19
bzr==2.7.0
camera-calibration==1.12.20
camera-calibration-parsers==1.11.12
catkin==0.7.6
catkin-pkg==0.3.5
catkin-pkg-modules==0.3.5
chardet==2.3.0
configobj==5.0.6
controller-manager==0.11.5
controller-manager-msgs==0.11.5
cryptography==1.2.3
cv-bridge==1.12.4
cycler==0.9.0
Cython==0.25.2
defer==1.0.6
defusedxml==0.4.1
diagnostic-analysis==1.9.0
diagnostic-common-diagnostics==1.9.0
diagnostic-updater==1.9.0
docutils==0.12
dynamic-reconfigure==1.5.48
ecdsa==0.13
empy==3.3.2
enum34==1.1.2
filelock==2.0.10
funcsigs==1.0.2
futures==3.0.5
gazebo-plugins==2.5.13
gazebo-ros==2.5.13
gencpp==0.5.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.8
gennodejs==2.0.1
genpy==0.6.5
gyp==0.1
html5lib==0.999
httplib2==0.9.1
idna==2.0
image-geometry==1.12.4
interactive-markers==1.11.3
ipaddress==1.0.16
Keras==2.1.3
keyring==7.3
laser-geometry==1.6.4
launchpadlib==1.10.3
lazr.restfulclient==0.13.4
lazr.uri==1.0.3
lxml==3.5.0
lz4==0.7.0
Markdown==2.6.11
matplotlib==1.5.1
mercurial==3.7.3
message-filters==1.12.7
mock==2.0.0
moveit-commander==0.9.8
moveit-ros-planning-interface==0.9.8
moveit-ros-visualization==0.9.8
mpi4py==1.3.1
mpmath==0.19
msgpack-python==0.4.6
netifaces==0.10.4
nose==1.3.7
numpy==1.11.0
oauth==1.0.1
PAM==0.4.2
pandas==0.22.0
paramiko==1.16.0
pbr==3.1.1
Pillow==3.1.2
pluginlib==1.10.5
protobuf==3.5.1
psutil==3.4.2
pyasn1==0.1.9
pyasn1-modules==0.0.7
PyAssimp==3.0.1264
pycrypto==2.6.1
pydot==1.0.29
Pygments==2.1
pygobject==3.20.0
pygpgme==0.3
PyOpenGL==3.0.2
pyOpenSSL==0.15.1
pyparsing==2.0.3
pyserial==3.0.1
pysqlite==2.7.0
Pyste==0.9.10
python-apt==1.1.0b1
python-dateutil==2.4.2
python-debian==0.1.27
python-pcl==0.2
python-qt-binding==0.3.2
python-snappy==0.5
pytz==2014.10
pyxdg==0.25
PyYAML==3.11
qt-dotgraph==0.3.4
qt-gui==0.3.4
qt-gui-cpp==0.3.4
qt-gui-py-common==0.3.4
resource-retriever==1.12.3
roman==2.0.0
rosbag==1.12.7
rosboost-cfg==1.13.5
rosclean==1.13.5
roscreate==1.13.5
rosdep==0.11.5
rosdistro==0.6.2
rosdistro-modules==0.6.2
rosgraph==1.12.7
rosinstall==0.7.8
roslaunch==1.12.7
roslib==1.13.5
roslint==0.11.0
roslz4==1.12.7
rosmake==1.13.5
rosmaster==1.12.7
rosmsg==1.12.7
rosnode==1.12.7
rosparam==1.12.7
rospkg==1.1.1
rospkg-modules==1.1.1
rospy==1.12.7
rosservice==1.12.7
rostest==1.12.7
rostopic==1.12.7
rosunit==1.13.5
roswtf==1.12.7
rqt-action==0.4.9
rqt-bag==0.4.8
rqt-bag-plugins==0.4.8
rqt-console==0.4.8
rqt-dep==0.4.8
rqt-graph==0.4.8
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.8
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.7
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.8
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.7
rqt-robot-steering==0.5.7
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.7
rqt-service-caller==0.4.8
rqt-shell==0.4.8
rqt-srv==0.4.8
rqt-tf-tree==0.5.7
rqt-top==0.4.8
rqt-topic==0.4.8
rqt-web==0.4.8
rviz==1.12.10
scipy==1.0.0
SecretStorage==2.1.3
sensor-msgs==1.12.5
service-identity==16.0.0
simplejson==3.8.1
six==1.10.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.7.19
srdfdom==0.4.2
sympy==1.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.0
tf==1.11.8
tf-conversions==1.11.8
tf2-geometry-msgs==0.5.15
tf2-kdl==0.5.15
tf2-py==0.5.15
tf2-ros==0.5.15
topic-tools==1.12.7
trollius==2.0.1
Twisted==16.0.0
txaio==1.0.0
urdfdom-py==0.3.3
vcstools==0.1.39
wadllib==1.3.2
Werkzeug==0.14.1
wstool==0.1.13
xacro==1.11.2
zope.interface==4.1.3
